test-eval/pyproject.toml
[project]
name = "test-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-evaluation>=1.12.0",
    "dotenv>=0.9.9",
    "graphrag>=2.7.0",
    "pandas>=2.3.3",
    "rich>=13.0.0",
    "dynaconf>=3.2.11",
]
